change mana from int to float + re-enable usage blocked predicate for breath ability + reduce mana cost for nether breath

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/common/codecs/ability/ActionContainer.java
@@ -46,7 +46,7 @@ public record ActionContainer(AbilityTargeting effect, LevelBasedValue triggerRa
         ManaCost.Type manaCostType = manaCost.map(ManaCost::type).orElse(null);
 
         if (manaCostType == ManaCost.Type.TICKING && (abilityType == DragonAbility.Type.PASSIVE || abilityType == DragonAbility.Type.ACTIVE_CHANNELED)) {
-            int cost = (int) manaCost.get().manaCost().calculate(instance.level()); // TODO :: remove cast to int
+            float cost = manaCost.get().manaCost().calculate(instance.level());
 
             if (ManaHandler.hasEnoughMana(dragon, cost)) {
                 ManaHandler.consumeMana(dragon, cost); // TODO :: make this return a boolean and remove 'hasEnoughMana'?
